docs: add api docs
