+++ packages/react-sketch-canvas/src/index.tsx
@@ -1,3 +1,8 @@
 export * from "./Canvas";
 export * from "./ReactSketchCanvas";
 export * from "./types";
+export { CanvasRef } from "./Canvas/types";
+export { CanvasProps } from "./Canvas/types";
+export { AllowOnlyPointerType } from "./Canvas/types";
+export { ReactSketchCanvasProps } from "./ReactSketchCanvas/types";
+export { ReactSketchCanvasRef } from "./ReactSketchCanvas/types";
